Fix lint issues and tests names
