+++ docs/whats_new.md
@@ -971,11 +971,7 @@ The adaptor middleware has been significantly optimized for performance and effi
 
 We are excited to introduce a new option in our caching middleware: Cache Invalidator. This feature provides greater control over cache management, allowing you to define a custom conditions for invalidating cache entries.
 Additionally, the caching middleware has been optimized to avoid caching non-cacheable status codes, as defined by the [HTTP standards](https://datatracker.ietf.org/doc/html/rfc7231#section-6.1). This improvement enhances cache accuracy and reduces unnecessary cache storage usage.
-
-Other highlights:
-
-- Cached responses now include an RFC-compliant `Age` header.
-- Response `Cache-Control` directives such as `max-age` and `no-store` are honored.
+Cached responses now include an RFC-compliant Age header, providing a standardized indication of how long a response has been stored in cache since it was originally generated. This enhancement improves HTTP compliance and facilitates better client-side caching strategies.
 
 ### CORS
